+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 - Added extra cheats that can be used in non sv_cheats server: '-giveability x' '-removeability x'
 - Added chat commands that can be used to disable some game mechanics: Disable the increasing respawn time mechanic (-enablerespawn or -er), disable the anti-kamikaze mechanic (-enablekamikaze or -ek), you can also use -ec to enable cheats.
 - Fixed global tusk kick and increased its kick distance
-- Temporarily blocked single player abilities in multiplayer games until better solution can be found
+- Temporarily blocked single player abilities in multiplayer VS games until better solution can be found. They can still be used if all human players are on the same team (coop).
 - Improved Cluster Rockets OP
 - Added Single Player OP versions of the following abilities: Dark Rift, Vacumn, Chaos Meteor and versions of OP BloodLust 
 - Increased Portal cast range