+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 6.9 - Development Version
+## 7.0 - Development Version
+- You can now use '-bot mode' chat command to see what mode bots are in (early game or late game). '-bot switch' can be used to alternative between modes, but can only be used when cheats are enable or in single player. 
+
+## 6.9 - Steam Version
 - You can enable cheats in a multiplayer games now if everbody types '-enablecheats'
 - Reduced the amount of health neutral power modifier gives per level (100 to 80)
 - Added extra cheats that can be used in non sv_cheats server: '-giveability x' '-removeability x'
@@ -13,7 +16,7 @@ All notable changes to this project will be documented in this file.
 - Added Single Player OP versions of the following abilities: Dark Rift, Vacumn, Chaos Meteor and versions of OP BloodLust 
 - Increased Portal cast range
 
-## 6.8 - Steam Version
+## 6.8 
 - Tomes now get lower cooldown on levelup
 - Magic Barrier decay reduced to 3 per second instead of 5, but no longer works with abilities with 0 cooldown (e.g. Arcane Orb)
 - You can now use the cheats '-gold', '-respawn' and '-refresh' in any single player game, with or without sv_cheats being activated.